# swagger.yaml
openapi: 3.0.0
info:
  title: P9 E Commerce Api
  contact: {}
  version: '1.0.0'
servers:
- url: https://e-commerce-api-yyqb.onrender.com
  variables: {}
paths:
...
  /orders:
    get:
      tags:
      - Orders
      summary: Get All Orders
      operationId: GetAllOrders
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
      deprecated: false
    post:
      tags:
      - Orders
      summary: Create Order
      operationId: CreateOrder
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateOrderRequest'
              - {}
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /orders/{id}:
    parameters:
        - name: id
          in: path
          description: 'order ID'
          required: true
          schema:
            type: string
    get:
      tags:
      - Orders
      summary: Get Single Order
      operationId: GetSingleOrder
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  tax: number
                  shippingFee: number
                  items:
                  - name: string
                    price: string
                    image: string
                    amount: number
                    product: string
              example:
                tax: number
                shippingFee: number
                items:
                - name: string
                  price: string
                  image: string
                  amount: number
                  product: string
      deprecated: false
    post:
      tags:
      - Orders
      summary: Update Order
      operationId: UpdateOrder
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateOrderRequest'
              - example:
                  paymentIntentId: string
            example:
              paymentIntentId: string
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
components:
  schemas:
...
    CreateOrderRequest:
      title: CreateOrderRequest
      required:
      - tax
      - shippingFee
      - items
      type: object
      properties:
        tax:
          type: integer
          format: int32
        shippingFee:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
          description: ''
      example:
        tax: 499
        shippingFee: 799
        items:
        - name: bed
          price: '2699'
          image: http://dummyimage.com/206x100.png/5fa2dd/ffffff
          amount: 3
          product: 6496a0145073c874cff6019b
    Item:
      title: Item
      required:
      - name
      - price
      - image
      - amount
      - product
      type: object
      properties:
        name:
          type: string
        price:
          type: string
        image:
          type: string
        amount:
          type: integer
          format: int32
        product:
          type: string
      example:
        name: bed
        price: '2699'
        image: http://dummyimage.com/206x100.png/5fa2dd/ffffff
        amount: 3
        product: 6496a0145073c874cff6019b
    UpdateOrderRequest:
      title: UpdateOrderRequest
      required:
      - paymentIntentId
      type: object
      properties:
        paymentIntentId:
          type: string
      example:
        paymentIntentId: randomValue
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...